Split of README
To keep the documentation close to the corresponding code,
the readme has been splitted. One part, describing the project in general,
remains in this project.
The other part, describing the installation procedure, is moved into the
landscape-setup  project, as it describes the usage of the scripts
located in that project.